+++ client/src/App.tsx
@@ -1,9 +1,8 @@
 import React from 'react';
-import { BrowserRouter as Router, Switch } from "react-router-dom";
+import { BrowserRouter as Router } from "react-router-dom";
 import Loading from './components/Loading/Loading';
 import { useAuth0 } from '@auth0/auth0-react';
 import Navbar from './components/Navbar/Navbar';
-import { makeStyles, createStyles, Theme } from "@material-ui/core/styles";
 import LoggedInLanding from "./views/Landing/loggedInLanding"
 import LoggedOutLanding from "./views/Landing/loggedOutLanding"
