+++ client/src/components/Landing/TeamBio.tsx
