+++ client/src/components/Landing/landing.tsx
@@ -3,7 +3,7 @@ import { useTheme } from '@material-ui/core/styles';
 import useMediaQuery from '@material-ui/core/useMediaQuery';
 // import AuthNav from "../../components/Login/Auth-nav";
 // import SignupButton from "../../components/Login/Signup-button"
-import AuthSignup from "../../components/Login/Auth-signup"
+import AuthSignup from "../Login/Auth-signup"
 
 import { Box, Container } from "@material-ui/core"
 import { makeStyles, createStyles, Theme } from "@material-ui/core/styles";
